BETA: testing char array buffers for response
attempting to optimize memory when dealing with response strings.

Previously buildJsonString returned a string object for the response

I experimented with using a char array to handle the same data.
During testing I observed some delays and device reboots. I am not sure
if this is related to network problems, a faulty device, or the code change

Next I will test previous stable branch on same device to rule out network and device